+++ CHANGES.md
@@ -34,6 +34,8 @@ Update version in extension.json.
 YYYY-MM-DD
 Speechoid: 0.1.2
 
+* [T402518](https://phabricator.wikimedia.org/T402518) Combine utterances to audio files
+
 ### 0.1.13
 2025-11-10
 Speechoid: 0.1.2